Refactor and extend report definition handling logic
Refactored report definition resources by separating paginated list and single definition handling into distinct endpoints. Enhanced handling of expired report entries in the frontend, ensuring proper fetching, formatting, and display. Updated test cases to validate the refactor and added cases for new functionality.
